+++ app/src/main/java/com/example/tourviet/verify_otp_recovery.java
@@ -1,11 +1,11 @@
 package com.example.tourviet;
 
 public class verify_otp_recovery {
-    int userId;
+    long userId;
     String newPassword;
     String verifyCode;
 
-    public verify_otp_recovery(int userId, String newPassword, String verifyCode) {
+    public verify_otp_recovery(Long userId, String newPassword, String verifyCode) {
         this.userId = userId;
         this.newPassword = newPassword;
         this.verifyCode = verifyCode;
